+++ app/domain/token_factory.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'date'
+
 class TokenFactory < Dry::Struct
 
   NoSigningKey = ::Util::ErrorClass.new(
@@ -10,9 +12,19 @@ class TokenFactory < Dry::Struct
   def signing_key(account)
     slosilo["authn:#{account}".to_sym] || raise(NoSigningKey, account)
   end
-    
+
+  def signed_host_key(account:, username:)
+    exp_t = Time.now + Rails.application.config.conjur_config.host_authorization_token_ttl.to_i
+    signing_key(account).issue_jwt(sub: username, exp: exp_t)
+  end
+
   def signed_token(account:, username:)
-    signing_key(account).issue_jwt(sub: username)
+    if username.starts_with?('host/')
+      signed_host_key(account: account, username: username)
+    else
+      exp_t = Time.now + Rails.application.config.conjur_config.user_authorization_token_ttl.to_i
+      signing_key(account).issue_jwt(sub: username, exp: exp_t)
+    end
   end
 
 end